+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Improve readability of README
 
+### Changed
+- Migrate to `aiohttp` with the support of passing a session
+- Disable socket use in tests
+
 ## [0.1.0] - 2022-08-22
 ### Added
 - Setup initial project structure (thanks Wolt!)